requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
av==11.0.0
certifi==2024.2.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
ctranslate2==4.0.0
decorator==4.4.2
distro==1.9.0
exceptiongroup==1.2.0
faster-whisper==1.0.1
ffmpeg-python==0.2.0
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.3
humanfriendly==10.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
numpy==1.26.4
onnxruntime==1.17.1
openai==0.28.0
packaging==23.2
pillow==10.2.0
proglog==0.1.10
protobuf==4.25.3
pydantic==2.6.3
pydantic_core==2.16.3
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.1
sympy==1.12
tokenizers==0.13.3
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
yarl==1.9.4
